# hydrators should not be proxy, now works gracefully with final classes