… src/GeneratedHydrator/ClassGenerator/Hydrator/PropertyGenerator/PropertyAccessor.php
… tests/GeneratedHydratorTest/ClassGenerator/Hydrator/PropertyGenerator/PropertyAccessorTest.php
… tests/GeneratedHydratorTestAsset/ClassWithPrivatePropertiesAndParent.php
… tests/GeneratedHydratorTestAsset/ClassWithPrivatePropertiesAndParents.php