By not extending the original class, we can hydrate final classes. It makes the generate code simpler, potentially faster for complex classes to hydrate, and generation code a lot simpler.

It uses the Closure::bind() mecanism for all properties, no matter their visibility, but it creates a single closure for each parent class instead of a closure per method, which drastically reduces the number of potential function calls when hydrating objects with a complex hierarchy and a lot of private properties.

If also fixes a bug where when attempting to hydrate objects partially (without all properties within the array) notices were thrown, now they are not anymore!

It also fixes #35 

And potentially #51 because Closures are ordered following the hierarchy: if a parent private property is hidden by a child class, the child will always get the private property, and not the parent.

It removes a few code for manipulating the AST, since it's not necessary anymore. Actually, we could probably remove ALL the AST manipulation dependency and just write a PHP template with the added Closures inside, which would make this code even faster and simpler.