@pounard 2017-01-11
It also adds tests for class hierarchy and values.