@flip111 2017-01-11
can you do some benchmarks on php 5 & 7 for this PR ?