@pounard 2017-01-11
I did with PHP7, with no private properties, I have a 40% penalty, with 2 private properties across a parent and a child class, I get 20%, so as an educated guess I would say that reaching 4 or 6 private properties, I'm on-par in term of performances, but I'll do some more tomorrow and give you the methodology and results.

If I'm not mistaken, this API doesn't work with PHP5 due to strict types usage.