@flip111 2017-01-11
i was using this library with php5 before and it was working then

not sure what you mean with 40% penalty and on-par performance. i mean to compare this pr with master. i will look at issue 60 later