@pounard 2017-01-11
40% penalty means that in ideal conditions (which never happens in real life environments) I am 40% slower, but I'm quite sure that with real life objects (much more complex data structures) there would be no significant difference.