@pounard 2017-01-11
@malukenho I mostly agree with your requested changes, but some do change the codestyle of the library, so I really want to wait for @Ocramius opinion, it's his call to make.