@Ocramius 2017-01-12
> with real life objects (much more complex data structures) there would be no significant difference.

The library was designed for inclusion in ORM v3, so any bit of performance we can squeeze out of it should be considered. If the hydrator codegen without inheritance is something that fixes handling of `final` and classes already implementing the hydrator interface, then this should probably be considered a fallback, rather than a replacement solution.