@Ocramius 2017-01-12
> can you do some benchmarks on php 5 & 7 for this PR ?

Indeed, we need to add benchmarks ([`phpbench`-based](https://github.com/phpbench/phpbench)) to the library.
PHP 5 is no longer the target audience though.