@pounard 2017-01-12
> The library was designed for inclusion in ORM v3, so any bit of performance we can squeeze out of it should be considered

It surely needs more benchmarks to tell, but I'd probably say that the performance impact is mostly due to isset() calls, and probably not by the fact there's closures. I'll test that.