@Ocramius 2017-01-12
> I'll test that.

The amount of opcodes being generated usually gives you a good idea of where the slowness is.