@pounard 2017-01-12
> The library was designed for inclusion in ORM v3

What are the requirements for ORM v3 regarding this library ?

> If the hydrator codegen without inheritance is something that fixes handling of final and classes already implementing the hydrator interface, then this should probably be considered a fallback, rather than a replacement solution.

Not even sure why inheritance here is even wanted to start with, I think that it should get rid of, there is no need at all, and it replaces your objects by proxies and it actually makes no sense to me, except if it is required by ORM v3.

I think it make the overall code of this library much harder to read and maintain since you have to parse the class AST, replace its methods, add some others, change its name, add the extension, etc etc etc... Generating much more simple hydrators could also make the AST completly useless in the end.

And I often read your blog posts, and watch from time to time your sessions, isn't composition and single responsibility supposed to be better than inheritance ? :) My guess here, especially in this case, inheritance doesn't serve any purpose except making the whole much more complex.

Could you elaborate why this should be a fallback instead of being the common case?

Could you also elaborate about why it's a good thing to keep a very complex generation mechanism while it could be very much simpler and serve the exact same features?

I guess you have a roadmap I don't know anything about, but right now I see plenty of code that could be overly simplified and still be on-par in term of features and performance.

Thank you for your time and answers, looking forward to read your explanations!