@pounard 2017-01-12
> The amount of opcodes being generated usually gives you a good idea of where the slowness is.

I'm doing it sometime, but nothing worth real figures, especially that opcode is itself an high-level IR and all operations do not cost the same, so less doesn't really mean faster. Except if you know pretty well how the zend internals are working, I guess it dangerous to analyse code performance only through this.