@Ocramius 2017-01-12
> What are the requirements for ORM v3 regarding this library ?

Hydration in ORM has a massive performance impact, specifically method call time (reflection). If we can reduce that time in any way by providing an abstraction that does it for us, then the abstraction should always strive for best performance there.

 > Not even sure why inheritance here is even wanted to start with, I think that it should get rid of, there is no need at all, and it replaces your objects by proxies and it actually makes no sense to me, except if it is required by ORM v3.

The reason why inheritance was that any scope change would need a closure method call (expensive). Another possible solution is to return a closure that is already bound to the correct scope, and getting rid of the hydrator interface overall.

 > I think it make the overall code of this library much harder to read and maintain since you have to parse the class AST, replace its methods, add some others, change its name, add the extension, etc etc etc... Generating much more simple hydrators could also make the AST completly useless in the end.

Agree - if we can get rid of it, yet retain the same performance, then that's a very good thing.

 > And I often read your blog posts, and watch from time to time your sessions, isn't composition and single responsibility supposed to be better than inheritance ? :) My guess here, especially in this case, inheritance doesn't serve any purpose except making the whole much more complex.

As mentioned above, inheritance is used here as a hack to get rid of a single method call, yet retaining compat with the hydrator interfaces.

 > Could you elaborate why this should be a fallback instead of being the common case?

If the new approach cannot provide the same performance as the previous one, then it should only be used when a `final` class is involved, or when a class implementing `hydrate` or `extract` is involved, or a class with a `final` constructor.