@Ocramius 2017-01-12
> Could you also elaborate about why it's a good thing to keep a very complex generation mechanism while it could be very much simpler and serve the exact same features?

The only feature provided by this library is speed. If that is not the absolute focus of the consumer of the library, then a simplistic, manually-written hydrator is sufficient in most scenarios.

 > I guess you have a roadmap I don't know anything about, but right now I see plenty of code that could be overly simplified and still be on-par in term of features and performance.

No, sorry, I don't have a roadmap for most OSS stuff, since I really work on it when I got time.
The on-par needs to be validated by a proper test suite there, which we should add separately, before eventually merging this patch.