@Ocramius 2017-01-12
> I'm doing it sometime, but nothing worth real figures, especially that opcode is itself an high-level IR and all operations do not cost the same, so less doesn't really mean faster. Except if you know pretty well how the zend internals are working, I guess it dangerous to analyse code performance only through this.

Yes, but at a quick glance you see what is going on :-)