@pounard 2017-01-13
Just for the posterity, did stable benchmarks, one with isset() and without the other().

WITH isset():
```
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+----------+---------+--------+------------+
| benchmark                          | subject                        | groups | params | revs | its | mem_peak   | best     | mean     | mode     | worst    | stdev   | rstdev | diff       |
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+----------+---------+--------+------------+
| InheritanceClassHydrationBench     | benchConsume                   |        | []     | 100  | 200 | 3,424,104b | 4.330μs  | 4.678μs  | 4.507μs  | 8.270μs  | 0.421μs | 9.00%  | +181.61%   |
| ZendHydrationBench                 | benchMethodAllPrivate          |        | []     | 100  | 200 | 1,584,032b | 35.160μs | 38.065μs | 36.524μs | 65.460μs | 4.051μs | 10.64% | +2,191.18% |
| ZendHydrationBench                 | benchMethodAllProtected        |        | []     | 100  | 200 | 1,584,032b | 35.150μs | 38.570μs | 36.727μs | 67.390μs | 4.501μs | 11.67% | +2,221.59% |
| ZendHydrationBench                 | benchMethodAllPublic           |        | []     | 100  | 200 | 1,584,032b | 34.910μs | 38.151μs | 36.553μs | 57.600μs | 3.769μs | 9.88%  | +2,196.39% |
| ZendHydrationBench                 | benchMethodInheritance         |        | []     | 100  | 200 | 1,584,032b | 35.020μs | 37.897μs | 36.499μs | 60.210μs | 3.373μs | 8.90%  | +2,181.08% |
| ZendHydrationBench                 | benchMethodInheritanceDeep     |        | []     | 100  | 200 | 1,584,040b | 35.230μs | 38.220μs | 36.607μs | 62.290μs | 3.931μs | 10.28% | +2,200.53% |
| ZendHydrationBench                 | benchMethodMixed               |        | []     | 100  | 200 | 1,584,032b | 34.730μs | 38.821μs | 36.607μs | 76.430μs | 6.223μs | 16.03% | +2,236.74% |
| ZendHydrationBench                 | benchReflectionAllPrivate      |        | []     | 100  | 200 | 1,234,552b | 34.820μs | 36.995μs | 36.043μs | 55.660μs | 2.860μs | 7.73%  | +2,126.82% |
| ZendHydrationBench                 | benchReflectionAllProtected    |        | []     | 100  | 200 | 1,234,552b | 34.600μs | 36.922μs | 35.866μs | 53.620μs | 3.214μs | 8.70%  | +2,122.39% |
| ZendHydrationBench                 | benchReflectionAllPublic       |        | []     | 100  | 200 | 1,234,552b | 34.370μs | 37.391μs | 35.833μs | 54.040μs | 4.111μs | 10.99% | +2,150.63% |
| ZendHydrationBench                 | benchReflectionInheritance     |        | []     | 100  | 200 | 1,234,552b | 43.260μs | 46.638μs | 45.119μs | 66.430μs | 4.048μs | 8.68%  | +2,707.25% |
| ZendHydrationBench                 | benchReflectionInheritanceDeep |        | []     | 100  | 200 | 1,234,552b | 52.490μs | 56.150μs | 54.296μs | 83.250μs | 4.761μs | 8.48%  | +3,279.81% |
| ZendHydrationBench                 | benchReflectionMixed           |        | []     | 100  | 200 | 1,234,544b | 34.510μs | 36.644μs | 35.617μs | 62.800μs | 2.987μs | 8.15%  | +2,105.67% |
| InheritanceDeepClassHydrationBench | benchConsume                   |        | []     | 100  | 200 | 3,436,792b | 6.220μs  | 6.639μs  | 6.514μs  | 12.760μs | 0.598μs | 9.01%  | +299.59%   |
| AllPrivateClassHydrationBench      | benchConsume                   |        | []     | 100  | 200 | 3,417,320b | 2.530μs  | 2.737μs  | 2.614μs  | 5.060μs  | 0.273μs | 9.98%  | +64.75%    |
| MixedClassHydrationBench           | benchConsume                   |        | []     | 100  | 200 | 3,416,816b | 2.550μs  | 2.740μs  | 2.621μs  | 4.830μs  | 0.282μs | 10.29% | +64.95%    |
| AllPublicClassHydrationBench       | benchConsume                   |        | []     | 100  | 200 | 3,413,520b | 1.530μs  | 1.661μs  | 1.578μs  | 2.560μs  | 0.169μs | 10.18% | 0.00%      |
| AllProtectedClassHydrationBench    | benchConsume                   |        | []     | 100  | 200 | 3,417,056b | 2.530μs  | 2.720μs  | 2.635μs  | 7.000μs  | 0.345μs | 12.68% | +63.72%    |
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+----------+---------+--------+------------+
```

WITHOUT isset():
```
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+-----------+---------+--------+------------+
| benchmark                          | subject                        | groups | params | revs | its | mem_peak   | best     | mean     | mode     | worst     | stdev   | rstdev | diff       |
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+-----------+---------+--------+------------+
| InheritanceClassHydrationBench     | benchConsume                   |        | []     | 100  | 200 | 3,420,368b | 3.810μs  | 4.056μs  | 3.918μs  | 7.480μs   | 0.388μs | 9.56%  | +169.82%   |
| ZendHydrationBench                 | benchMethodAllPrivate          |        | []     | 100  | 200 | 1,584,032b | 35.230μs | 38.052μs | 36.703μs | 77.060μs  | 4.594μs | 12.07% | +2,431.20% |
| ZendHydrationBench                 | benchMethodAllProtected        |        | []     | 100  | 200 | 1,584,032b | 35.180μs | 37.996μs | 36.743μs | 65.900μs  | 3.945μs | 10.38% | +2,427.47% |
| ZendHydrationBench                 | benchMethodAllPublic           |        | []     | 100  | 200 | 1,584,032b | 35.060μs | 37.758μs | 36.377μs | 59.100μs  | 3.679μs | 9.74%  | +2,411.65% |
| ZendHydrationBench                 | benchMethodInheritance         |        | []     | 100  | 200 | 1,584,032b | 34.910μs | 38.034μs | 36.492μs | 60.980μs  | 4.060μs | 10.68% | +2,430.03% |
| ZendHydrationBench                 | benchMethodInheritanceDeep     |        | []     | 100  | 200 | 1,584,040b | 35.040μs | 38.058μs | 36.592μs | 53.070μs  | 3.573μs | 9.39%  | +2,431.65% |
| ZendHydrationBench                 | benchMethodMixed               |        | []     | 100  | 200 | 1,584,032b | 34.970μs | 38.032μs | 36.492μs | 56.570μs  | 3.744μs | 9.84%  | +2,429.93% |
| ZendHydrationBench                 | benchReflectionAllPrivate      |        | []     | 100  | 200 | 1,234,552b | 34.690μs | 37.012μs | 36.062μs | 50.270μs  | 2.651μs | 7.16%  | +2,362.04% |
| ZendHydrationBench                 | benchReflectionAllProtected    |        | []     | 100  | 200 | 1,234,552b | 34.520μs | 36.810μs | 35.766μs | 51.280μs  | 2.791μs | 7.58%  | +2,348.59% |
| ZendHydrationBench                 | benchReflectionAllPublic       |        | []     | 100  | 200 | 1,234,552b | 34.440μs | 36.896μs | 35.781μs | 52.470μs  | 3.072μs | 8.33%  | +2,354.30% |
| ZendHydrationBench                 | benchReflectionInheritance     |        | []     | 100  | 200 | 1,234,552b | 43.460μs | 46.503μs | 45.104μs | 63.950μs  | 3.487μs | 7.50%  | +2,993.38% |
| ZendHydrationBench                 | benchReflectionInheritanceDeep |        | []     | 100  | 200 | 1,234,552b | 52.550μs | 56.611μs | 54.311μs | 116.840μs | 6.237μs | 11.02% | +3,665.80% |
| ZendHydrationBench                 | benchReflectionMixed           |        | []     | 100  | 200 | 1,234,544b | 34.520μs | 37.423μs | 35.858μs | 68.700μs  | 4.264μs | 11.39% | +2,389.37% |
| InheritanceDeepClassHydrationBench | benchConsume                   |        | []     | 100  | 200 | 3,428,128b | 5.410μs  | 5.832μs  | 5.616μs  | 11.990μs  | 0.714μs | 12.25% | +287.92%   |
| AllPrivateClassHydrationBench      | benchConsume                   |        | []     | 100  | 200 | 3,414,512b | 2.360μs  | 2.579μs  | 2.450μs  | 5.070μs   | 0.386μs | 14.97% | +71.57%    |
| MixedClassHydrationBench           | benchConsume                   |        | []     | 100  | 200 | 3,414,152b | 2.370μs  | 2.555μs  | 2.432μs  | 4.490μs   | 0.246μs | 9.63%  | +69.95%    |
| AllPublicClassHydrationBench       | benchConsume                   |        | []     | 100  | 200 | 3,411,784b | 1.380μs  | 1.503μs  | 1.426μs  | 2.700μs   | 0.176μs | 11.71% | 0.00%      |
| AllProtectedClassHydrationBench    | benchConsume                   |        | []     | 100  | 200 | 3,414,248b | 2.350μs  | 2.534μs  | 2.432μs  | 4.670μs   | 0.227μs | 8.98%  | +68.55%    |
+------------------------------------+--------------------------------+--------+--------+------+-----+------------+----------+----------+----------+-----------+---------+--------+------------+
```